removed unused postcss plugins; refactored landing page styles
