+++ packages/client/package.json
@@ -47,11 +47,9 @@
     "npm-run-all": "4.1.5",
     "postcss": "^8.4.21",
     "postcss-cli": "^10.1.0",
-    "postcss-color-mod-function": "3.0.3",
     "postcss-extend": "1.0.5",
     "postcss-for": "2.1.1",
     "postcss-import": "^15.1.0",
-    "postcss-mixins": "9.0.4",
     "postcss-nesting": "^11.2.2",
     "postcss-preset-env": "^8.1.0",
     "prettier": "^2.7.1",
