+++ packages/client/postcss.config.js
@@ -1,9 +1,7 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import cssnano from 'cssnano';
-import postcssColorMod from 'postcss-color-mod-function';
 import postcssExtend from 'postcss-extend';
 import postcssFor from 'postcss-for';
-import postcssMixins from 'postcss-mixins';
 import postcssPresetEnv from 'postcss-preset-env';
 
 const cssnanoPlugin = cssnano({
@@ -11,13 +9,7 @@ const cssnanoPlugin = cssnano({
 });
 const postcssPresetEnvPlugin = postcssPresetEnv({ stage: 1 });
 
-const plugins = [
-  postcssExtend,
-  postcssMixins,
-  postcssFor,
-  postcssColorMod,
-  postcssPresetEnvPlugin,
-];
+const plugins = [postcssExtend, postcssFor, postcssPresetEnvPlugin];
 if (process.env.BUILD_MODE !== 'development') {
   plugins.push(cssnanoPlugin);
 }
