+++ packages/client/src/components/LandingPage/Header/Header.css
@@ -1,47 +1,45 @@
-.landing-page {
-  .header {
-    width: 100%;
-    height: 15vh;
+.landing-page-header {
+  width: 100%;
+  height: 15vh;
+}
+
+.landing-seo-title {
+  position: fixed;
+  z-index: -2;
+}
+
+.landing-title-section {
+  padding-top: 3vh;
+}
+
+.landing-title-svg {
+  display: block;
+
+  width: 6em;
+  max-height: 12vh;
+  margin: 0 auto;
+
+  font-size: min(4em, 12vh);
+  font-weight: 500;
+}
+
+.landing-title-svg__text-copy {
+  fill: none;
+
+  &:nth-child(1) {
+    stroke: var(--game-title-primary-color);
+    stroke-width: 2px;
   }
 
-  .page-main-header {
-    position: fixed;
-    z-index: -2;
-  }
-
-  .game-title-section {
-    padding-top: 3vh;
-  }
-
-  .game-title-svg {
-    display: block;
-
-    width: 6em;
-    max-height: 12vh;
-    margin: 0 auto;
-
-    font-size: min(4em, 12vh);
-    font-weight: 500;
-
-    .game-title-svg__text-copy {
-      fill: none;
-
-      &:nth-child(1) {
-        stroke: var(--game-title-primary-color);
-        stroke-width: 2px;
-      }
-
-      &:nth-child(2) {
-        stroke: var(--game-ttile-secondary-color);
-        stroke-dasharray: 0.5em 2em;
-        stroke-width: 1px;
-        animation: landing-page-game-title-animation 60s infinite linear;
-      }
-    }
+  &:nth-child(2) {
+    stroke: var(--game-ttile-secondary-color);
+    stroke-dasharray: 0.5em 2em;
+    stroke-width: 1px;
+    animation: landing-title-animation 60s infinite linear;
   }
 }
 
-@keyframes landing-page-game-title-animation {
+@keyframes landing-title-animation {
   from {
     stroke-dashoffset: 0;
   }
