+++ packages/client/src/components/LandingPage/Main/Main.tsx
@@ -1,13 +1,13 @@
 import { Link } from 'react-router-dom';
-import gameDescriptionText from './game-description-text';
+import gameDescriptionText from './game-description';
 import './Main.css';
 
 export default function () {
   return (
-    <main className="main">
-      <section className="game-description">
-        <p className="game-description__text">{gameDescriptionText}</p>
-        <Link to="/game" className="game-description__join-link">
+    <main className="landing-page-main">
+      <section className="landing-game-description">
+        <p className="landing-game-description__text">{gameDescriptionText}</p>
+        <Link to="/game" className="landing-game-description__join-link">
           <h2>Играйте сейчас и узнайте!</h2>
         </Link>
       </section>
