+++ packages/client/src/components/LandingPage/Main/game-description.ts
