+++ packages/client/src/components/LandingPage/_Component/LandingPage.tsx
@@ -1,10 +1,10 @@
-import PageFooter from './Footer/_Component';
-import PageHeader from './Header/Header';
-import PageMain from './Main/Main';
-import PageWallpaper from './Wallpaper/Wallpaper';
+import PageFooter from '../Footer/_Component';
+import PageHeader from '../Header/Header';
+import PageMain from '../Main/Main';
+import PageWallpaper from '../Wallpaper/Wallpaper';
 import './LandingPage.css';
 
-export default function () {
+export function LandingPage() {
   return (
     <div className="landing-page">
       <PageWallpaper></PageWallpaper>
