+++ packages/client/src/components/LandingPage/_Component/index.ts
@@ -1,0 +1,1 @@
+export { LandingPage as default } from './LandingPage';
